FEATURE: a change in Product

+++ Model_ViewModel/Models/Product.cs
@@ -20,7 +20,7 @@ namespace Model_ViewModel.Models
         public Category Category { get; set; }
         public Person Person { get; set; }
         public int PersonID { get; set; }
-        public long CategoryID { get; set; }
+        public int CategoryID { get; set; }
         public virtual ICollection<ProductImages> ProductImages { get; set; }
         public virtual ICollection<ProductFeatures> ProductFeatures { get; set; }
